[INTERNAL] move range concepts to std module
