+++ include/seqan3/core/metafunction/range.hpp
@@ -48,7 +48,7 @@
 #include <seqan3/core/metafunction/pre.hpp>
 #include <seqan3/core/metafunction/basic.hpp>
 #include <seqan3/core/metafunction/iterator.hpp>
-#include <seqan3/range/concept.hpp>
+#include <seqan3/std/concept/range.hpp>
 #include <seqan3/std/concept/iterator.hpp>
 
 // TODO(h-2): add innermost_reference instead of or addition to innermost_value_type?
