+++ include/seqan3/range/all.hpp
@@ -39,7 +39,6 @@
 
 #pragma once
 
-#include <seqan3/range/concept.hpp>
 #include <seqan3/range/container/all.hpp>
 #include <seqan3/range/decorator/all.hpp>
 #include <seqan3/range/view/all.hpp>
