+++ include/seqan3/range/detail/random_access_iterator.hpp
@@ -45,7 +45,7 @@
 #include <range/v3/utility/iterator_traits.hpp>
 #include <range/v3/range_traits.hpp>
 
-#include <seqan3/range/concept.hpp>
+#include <seqan3/std/concept/range.hpp>
 #include <seqan3/std/concept/iterator.hpp>
 
 namespace seqan3::detail
