+++ include/seqan3/range/view/all.hpp
@@ -40,7 +40,6 @@
 #pragma once
 
 #include <seqan3/range/view/char_to.hpp>
-#include <seqan3/range/view/concept.hpp>
 #include <seqan3/range/view/complement.hpp>
 #include <seqan3/range/view/convert.hpp>
 #include <seqan3/range/view/deep.hpp>
