+++ include/seqan3/range/view/detail.hpp
@@ -45,7 +45,7 @@
 
 #include <seqan3/core/metafunction/range.hpp>
 #include <seqan3/core/metafunction/template_inspection.hpp>
-#include <seqan3/range/concept.hpp>
+#include <seqan3/std/concept/range.hpp>
 
 namespace seqan3::detail
 {
