+++ include/seqan3/range/view/trim.hpp
@@ -42,7 +42,7 @@
 #include <range/v3/view/take_while.hpp>
 
 #include <seqan3/alphabet/quality/quality_composition.hpp>
-#include <seqan3/range/concept.hpp>
+#include <seqan3/std/concept/range.hpp>
 #include <seqan3/range/view/deep.hpp>
 
 namespace seqan3::detail
