+++ test/unit/range/view/view_to_rank_test.cpp
@@ -39,7 +39,7 @@
 #include <range/v3/view/reverse.hpp>
 
 #include <seqan3/alphabet/nucleotide/dna5.hpp>
-#include <seqan3/range/view/concept.hpp>
+#include <seqan3/std/concept/range.hpp>
 #include <seqan3/range/view/to_rank.hpp>
 
 using namespace seqan3;
