+++ test/unit/range/view/view_translation_test.cpp
@@ -43,8 +43,8 @@
 #include <seqan3/core/detail/reflection.hpp>
 #include <seqan3/range/container/concatenated_sequences.hpp>
 #include <seqan3/range/view/complement.hpp>
-#include <seqan3/range/view/concept.hpp>
 #include <seqan3/range/view/translation.hpp>
+#include <seqan3/std/concept/range.hpp>
 
 #include <range/v3/view/take.hpp>
 #include <range/v3/algorithm/equal.hpp>
